requirements.txt
Django==3.2.23
djangorestframework
psycopg2-binary
requests
folium
geopy
gunicorn